🐛 fix : 경로 수정
