+++ frontend/linki-user/src/api/proposal.js
@@ -38,7 +38,7 @@ export const proposalAPI = {
   // 제안서 수정
   updateProposal: async (proposalId, proposalData) => {
     try {
-      const response = await httpClient.post(`/v1/api/influencer/proposals/${proposalId}`, proposalData);
+      const response = await httpClient.put(`/v1/api/influencer/proposals/${proposalId}`, proposalData);
       return response.data;
     } catch (error) {
       console.error('Failed to update proposal:', error);
@@ -58,26 +58,42 @@ export const proposalAPI = {
   },
 
   // 캠페인별 제안서 목록 조회
-  getCampaignProposals: async (campaignId, params = {}) => {
+  // getCampaignProposals: async (campaignId, params = {}) => {
+  //   try {
+  //     const response = await httpClient.get(`/v1/api/influencer/campaigns/${campaignId}/proposals`, {
+  //       params: {
+  //         _page: params._page || 1,
+  //         _limit: params._limit || 10,
+  //         _sort: params._sort || 'createdAt',
+  //         _order: params._order || 'desc'
+  //       }
+  //     });
+  //     return {
+  //       proposals: response.data,
+  //       totalItems: parseInt(response.headers['x-total-count'] || '0')
+  //     };
+  //   } catch (error) {
+  //     console.error('Failed to fetch campaign proposals:', error);
+  //     return {
+  //       proposals: [],
+  //       totalItems: 0
+  //     };
+  //   }
+  // },
+   // 제안서 제출
+   submitProposal: async (campaignId, contents) => {
     try {
-      const response = await httpClient.get(`/v1/api/influencer/campaigns/${campaignId}/proposals`, {
-        params: {
-          _page: params._page || 1,
-          _limit: params._limit || 10,
-          _sort: params._sort || 'createdAt',
-          _order: params._order || 'desc'
-        }
-      });
-      return {
-        proposals: response.data,
-        totalItems: parseInt(response.headers['x-total-count'] || '0')
-      };
+      const response = await httpClient.post(`/v1/api/influencer/campaigns/${campaignId}/proposals`, {
+        contents: contents,
+        campaign_id: String(campaignId),
+        status: 'PENDING',
+        submitted_at: new Date().toISOString(),
+        influencer_id: 'INF001' // TODO: 실제 로그인한 사용자 ID로 대체
+      })
+      return response.data
     } catch (error) {
-      console.error('Failed to fetch campaign proposals:', error);
-      return {
-        proposals: [],
-        totalItems: 0
-      };
+      console.error('Error submitting proposal:', error)
+      throw error
     }
-  }
+  },
 }
